Changes in UI -> home login signup page
